third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/location-of-the-media-resource/currentSrc-expected.txt

This is a testharness.js-based test.
[FAIL] audio.currentSrc after adding source element with src attribute "data:,"
  assert_equals: expected "data:," but got ""
[FAIL] video.currentSrc after adding source element with src attribute "data:,"
  assert_equals: expected "data:," but got ""
Harness: the test ran to completion.
